requirements.txt
alpaca-trade-api>=3.0.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
websocket-client>=1.6.0
pytz>=2023.3
python-dateutil>=2.8.0
lxml>=4.9.0
backtrader>=1.9.0
streamlit>=1.20.0
plotly>=5.0.0
torch>=2.0.0
torchvision>=0.15.0
